refactor(hodo-list): Rename useBoolean file
